ingress: Add a separate load balancer for the Docker registry.
This works around a "fun" issue with DO load balancers.  Apparently they are OK with you sending
traffic from the internal network to the public IP... but they no longer add the proxy protocol
header.  This means that when Envoy sees the connection, it closes it because the header doesn't
parse.

The workaround here is to have a separate load balancer for anything that needs to be reachable from
the internal network, with the proxy protocol disabled.  We give up learning the remote IP address,
though, so it's not optimal.

Why do we even need to make internal requests?  Can't we just add "registry.jrock.us" as an alias to
"registry.docker-registry.svc.cluster.local" in CoreDNS?  Yes, we can, and that works.  But, the
registry image pull runs outside of a Kubernetes pod on the raw node, and the raw nodes do not use
the in-cluster DNS.  So we have no way of serving a split-horizon DNS that would be necessary to
make this work.  (And, TLS would be missing, because Envoy adds that.  I guess we could configure
the registry itself to serve over TLS, though, so if nodes were using in-cluster DNS, that's the
solution I would have chosen.)