ingress: add a separate envoy instance to handle "noproxy" resources
This is necessary because the main Envoy instance uses my custom container, which is pulled from my
own container registry, which is behind Envoy.  The problem is that there is quite the bootstrapping
problem here; if all Envoy instances go down, then no Envoy instance can pull the container to
restart.  This is really only a problem when nodes are added or removed from the cluster, but it's a
big problem.  This fixes the problem by using an externally-hosted Envoy container for traffic that
needs to pull containers.  Thus, this service can startup without any cluster resources, and then
the main public Envoy can start by pulling containers from my self-hosted registry.

This is all necessary because Digital Ocean doesn't let you configure node DNS for container pulls.